Fix no args call for newly created commands
Being 'foo' a non existing command, registering 'foo' and patching
'bar' to it would make `foo bar` work just fine.

But the no argument call (`foo`) would cause an error because the
logic was running the original command for the catch all clause,
but there is no original command for 'foo', it was just created.

This fix checks if the command being registered is an existing command.