Allow incremental patch
It is finally here!!
Now we can patch a bunch of commands without breaking the code.

This commit basically insert the new patch command function bellow the
shebang and a respective case for this new command bellow the switch
case.

For this, the mon register creates a minimal wrapper with a default code
only, instead of the empty one like before.

In summary, we can now do:
$ mon register ls
$ mon patch ls foo "echo 'fooooo'"
$ mon patch ls bar "echo 'baaaar'"